Add WalletService test in separate file
Delete chainId from message

Separate TransactionService test

MultiSignatureExecute to separate file
